chat.tmux.conf: Add help window for restricted mode, use neww -S
Provide a hotkey to browser the manual in its own window.

After input, nicm (tmux upstream) added "-S" to tmux(1) such that
the "new-window" command (in combination with "-d") first looks
for the given window name and selects the window if it exists
instead of trying to create a window that already exists.

Given that this makes chat.tmux.conf idempotent, we can now also reload
it at runtime to refresh settings.